src/requirements.txt
onnxruntime==1.20.1
onnxruntime-gpu==1.20.1
Pillow==11.1.0
torch==2.4
torchvision==0.19
fastapi==0.115.6
python-multipart==0.0.20
uvicorn==0.34.0